requirements.txt
streamlit
pypdf
faiss-cpu
openai
numpy
python-dotenv